Move MutuallyExclusiveOptions to the settings package
